requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
yt-dlp==2024.11.18
youtube-transcript-api==0.6.3
requests==2.32.3
